W252: the correct-size mid-band — +3 MATCH, and the overshoot is a signal, not a wall
**6604 -> 6607 first-party MATCH (80.3%).** 0 lost, 0 new over-window, 0 regressions.

## The population

287 functions were at nd 101-500 while already exactly the right size. That combination means the
object holds **no missing code and no invented code** — every differing word is a shape, lifetime or
type defect in source that is otherwise complete. Waves 249-250 had worked the low-nd end and measured
it out as floor; wave 246 had worked the prologue-divergent subset and repaired seven frames. This wave
applied both methods to the whole band.

**Three MATCHes, each from a different mechanism:**

- **`btlFormation 002bf690` nd175 -> 0.** The descending if-chain became an **ascending switch** on
  cases 1, 2, 3, 5, 6, 12; a target-local `u16` ABI alias removed an invented sign extension; and
  consuming the aliased `u32` predicate result directly removed the final `andi`. 444/448.
- **`mt_scene 003b58c0` nd111 -> 0.** An early unequal-id guard plus storing the call result directly,
  which moved the frame from 0x10 to retail's 0x20.
- **`fclShopMisc 003fd140` nd0**, completing the u8-contract-alias line of work that took it from
  nd306 in wave 246.

Largest diff wins: **`mdlFile 0034b1a0` nd243 -> 77, `0034b6b0` nd244 -> 79, `0034b3a0` nd221 -> 95,
`0034b8b0` nd222 -> 97** — one defect across four adjacent siblings; `fcl_misc 003c9850` nd297 -> 201;
`mdlFile 0034f6b0` nd122 -> 58; `y_runtime 0044a420` nd124 -> 120 by correcting a declared return type
from `u8` to `char`, which removed a retail-absent trailing `andi`; `datCalc 0030c0c0` nd108 -> 104 via
a `u32` -> `u16` field correction.

## The structural insight this wave produced

Agents kept finding **correct** lifetime fixes and reverting them for overshooting the window:
`datCalendar 00183be0` (+8 after the prologue divergence was fixed), `00186a40` (+8 after recovering
retail's saved-FPR lifetime), `mt_evtMain 00361270` (+16), `00367230` (+16), `datCalc 002ffdf0` (+24),
`btlCamera 002b8990` (+8 as a switch).

Reverting is correct procedure but the wrong conclusion, and I broadcast the reframe mid-wave:

**Every target here starts at exactly the right size. So if a lifetime fix is genuinely correct and
makes the object bigger, then by arithmetic there is invented code elsewhere in the same function.**
The overshoot proves you have found half the defect, not that the fix is wrong. Retail fits the logic
in that many bytes; if ours needs more, something else is surplus.

That reframe immediately produced a sharper class of result. W252Dat went back to `00183be0`, reapplied
the fix, and **located the exact 12-byte compensating saving** (a `default:` assignment instead of a
return) reaching 892/896 with the prologue repaired — then found the combined form regressed nd 353 ->
463 from pervasive register colouring, and reverted with that stated precisely. **A decline that names
the fix, the byte count and the blocker is worth far more than "floor"** — it tells the next wave
exactly what to look for.

## Residual character of the band

The declines cluster into three groups now clearly separated:

1. **Frame defects where the redundant value is unidentified** — `btlFormation 002bd8e0` (ours
   0x80/s0-s4 versus retail 0x70/s0-s3), `002be390` (0x90/s0-s6 versus 0x70/s0-s5), `002bea80` (same
   save set, 0x10 excess stack), `mdlFile 003226b0`/`00323240`/`003233a0`, `y_runtime 00431880`
   (0x50/s0-s3 versus 0x40/s0-s2), `00454d00` (retail copies the input vector through 0x20 more stack).
   Each is a real defect with a precisely quantified excess.
2. **Correct fixes blocked by an unlocated compensating saving** — quantified above at +8 to +24 bytes.
3. **Measured floors**, dominated by cyclic saved-register rotation with matching frame and save count
   (`mdlFile 00323e10` across three nested loop counters, `y_runtime 00452f70`) and by call-argument
   setup order.

The correct-size mid-band is now 277 functions, and the ones that remain are characterised rather than
merely unexplained.

+++ src/Main/Battle/Data/datCalc.c
@@ -167,8 +167,8 @@ u16 datCalcGetMaxHp(DatUnit* unit)
     {
         bonus -= (baseHp * 50) / 100;
     }
-
     hp += bonus;
+
     hp += (u16)datCalcCountEquipmentWithEffect(unit, 0x1a) * 10;
     hp += (u16)datCalcCountEquipmentWithEffect(unit, 0x1b) * 20;
     hp += (u16)datCalcCountEquipmentWithEffect(unit, 0x1c) * 30;
@@ -6480,7 +6480,7 @@ short FUN_0030bde0_narrow(u16 *param_1,s32 *param_2)
 int FUN_0030c0c0(void)
 
 {
-  u32 currentValue;
+  u16 currentValue;
   u16 current;
   u16 count;
   u16 inner;
